service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.12.0
aliyun oss sdk==2.6.0
httpx==0.27.0
pillow>=10.3.0
dashscope>=1.20.0
aiohttp==3.9.5
langchain==0.2.16
langchain-openai==0.1.25
langgraph==0.2.40
